pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
firebase-admin = "^6.6.0"
python-escpos = "^3.1"
pyserial = "^3.5"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
